requirements.txt
pandas>=2.2,<3.0
scikit-learn>=1.5,<2.0
streamlit>=1.45,<2.0
plotly>=5.24,<6.0
Pillow>=10.4,<12.0

